requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

cffi==1.15.0
    # via
    #   cmarkwrapper (setup.py)
    #   paka-cmark
paka-cmark==2.3.0
    # via cmarkwrapper (setup.py)
pycparser==2.21
    # via cffi
typing-extensions==4.1.1 ; python_version < "3.8"
    # via cmarkwrapper (setup.py)
